requirements.txt
ipython==8.17.2
ipywidgets==8.1.1
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
scikit_learn==1.3.2
torch==2.1.1
git+https://github.com/openai/CLIP.git
